Draft: Convert optimizely module to ts
